created renderWithProviderAndRouter file for jsx error ts migration
