+++ frontend/src/utils/test-utils.ts
@@ -18,36 +18,17 @@ import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import { setupStore } from "../store/configureStore";
 import { ToastContainer } from 'react-toastify';
+import { ReactNode } from "react";
 
-export const renderWithProviderAndRouter = (
-  ui,
-  {
-    preloadedState = {},
-    store = setupStore(preloadedState),
-    ...renderOptions
-  } = {}
-) => {
-  function Wrapper({ children }) {
-    return (
-      <Provider store={store}>
-        <BrowserRouter>
-          <ToastContainer position="top-center" hideProgressBar={true} autoClose={7000} theme='dark' />
-          {children}
-        </BrowserRouter>
-      </Provider>
-    );
-  }
 
-  return { store, ...render(ui, { wrapper: Wrapper, ...renderOptions }) };
-};
 
 
 export const createCategory = async () => {
   const createCategoryButton = await screen.findByRole('button', { name: /create/i})
   fireEvent.click(createCategoryButton)
   expect(await screen.findByText(/Create a new category/i)).toBeInTheDocument();
-  const button = screen.getByRole('button', {name: "Create"})
-  const input = screen.getByRole('textbox', {name: "New category name"})
+  const button = screen.getByRole<HTMLButtonElement>('button', {name: "Create"})
+  const input = screen.getByRole<HTMLTextAreaElement>('textbox', {name: "New category name"})
   fireEvent.change(input, {target: {value: "test_category"}})
   expect(input.value).toBe('test_category')
   expect(button).not.toBeDisabled()
